Don't combine reg_number and id
Initially, for some crazy reason, I assumed that registration number and id
would always be identical, but didn't actually implement anything to make this so.
This meant that creating an new object record failed since id wasn't auto generated. The two fields were also used interchangeably in some places. It was also going to be impossible to renumber objects.